Add validation tests from `validateJenkinsJobDsl` task
Although it is quite simple and elegant, this approach does not seem to
work as it is difficult to set the classpath from the `jenkins`
configuration.

This commit simply saves this step so that we can branch off this idea
and come back to it later if other ideas do not work.